extract core of webapp separate from shuttle.rs usage (e.g. just Router)
