+++ TODO.md
@@ -25,6 +25,6 @@
   - (/) simple "hello world" axum project working locally
   - (x) building locally in docker (podman)
   - (/) building and running remotely on fly.io
-  - (x) extract core of webapp separate from shuttle.rs usage (e.g. just Router)
+  - (/) extract core of webapp separate from shuttle.rs usage (e.g. just Router)
   - (x) use core in a fly.io shell, but with different secrets to distinguish usage
   - (x) leave deployed side-by-side in both for a day or so
