+++ backend/webapp/Cargo.toml
@@ -1,13 +1,10 @@
 [package]
-name = "fosdem2024"
+name = "webapp"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
 axum = { version = "0.6", features = ["macros"] }
-shuttle-axum = "0.35.0"
-shuttle-runtime = "0.35.0"
-shuttle-secrets = "0.35.2"
 tokio = "1.28.2"
 tracing = "0.1.40"
 axum-valid = "0.11"
@@ -15,4 +12,3 @@ validator = { version = "0.16", features = ["derive"] }
 askama = "0.12"
 serde = { workspace = true }
 shared = { path = "../shared" }
-webapp = { path = "../webapp" }
