+++ backend/webapp/src/router.rs
@@ -1,0 +1,84 @@
+use std::sync::Arc;
+
+use askama::Template;
+use axum::{
+    debug_handler,
+    extract::{Query, State},
+    response::Html,
+    routing::get,
+    Router,
+};
+use axum_valid::Valid;
+use serde::Deserialize;
+use shared::queryable::{Queryable, SearchItem};
+use validator::Validate;
+
+#[derive(Clone, Debug)]
+struct AppState {
+    queryable: Arc<Queryable>,
+}
+
+#[derive(Deserialize, Validate, Debug)]
+struct Params {
+    #[validate(length(min = 3, max = 100))]
+    q: String,
+    #[validate(range(min = 1, max = 20))]
+    limit: u8,
+    related: Option<String>,
+}
+
+#[derive(Template, Debug)]
+#[template(path = "search.html")]
+struct SearchTemplate {
+    query: String,
+    items: Vec<SearchItem>,
+}
+
+#[derive(Template)]
+#[template(path = "index.html")]
+struct IndexTemplate {}
+
+async fn index() -> Html<String> {
+    let page = IndexTemplate {};
+    let html = page.render().unwrap();
+    Html(html)
+}
+
+#[debug_handler]
+async fn search(
+    State(state): State<AppState>,
+    Valid(Query(params)): Valid<Query<Params>>,
+) -> axum::response::Result<Html<String>> {
+    match state
+        .queryable
+        .search(&params.q, params.limit, params.related.is_some())
+        .await
+    {
+        Ok(items) => {
+            let page = SearchTemplate {
+                query: params.q,
+                items,
+            };
+            let html = page.render().unwrap();
+            Ok(Html(html))
+        }
+        Err(_) => Err("search failed".into()),
+    }
+}
+
+pub async fn router(openai_api_key: &str, db_host: &str, db_key: &str) -> Router {
+    let state = AppState {
+        queryable: Arc::new(
+            Queryable::connect(&db_host, &db_key, &openai_api_key)
+                .await
+                .unwrap(),
+        ),
+    };
+
+    let router = Router::new()
+        .route("/", get(index))
+        .route("/search", get(search))
+        .with_state(state);
+
+    router
+}
